fix: ensure that due_items returned are always up-to-date
Previously, as a form of premature optimization, the `_navigate_to` method checked whether the new URL is different from the old URL before opening the new URL. This has caused repeated invocations of the `due_items` property method to return stale data.

The new implementation of `_navigate_to` does not do this optimization and returns the response object, so that it can also be used from the `reviews` property method.